+++ analysis/nearline/lisa/LisaTraceAnalysis.cxx
@@ -176,11 +176,11 @@ void LisaTraceAnalysis::Exec(Option_t* option)
 
         // ::::: Retrieve Data :::::
         int ch_ID = lisaAnaItem.Get_channel_id_traces();
-        double energy_febex = lisaAnaItem.Get_channel_energy();
-        double energy_MWD = lisaAnaItem.Get_channel_energy_MWD();
-        std::vector<int16_t> trace_febex = lisaAnaItem.Get_trace_febex();
-        std::vector<int16_t> trace_febex_0 = lisaAnaItem.Get_trace_febex_0();
-        std::vector<int16_t> trace_MWD = lisaAnaItem.Get_trace_MWD();
+        float energy_febex = lisaAnaItem.Get_channel_energy();
+        float energy_MWD = lisaAnaItem.Get_channel_energy_MWD();
+        std::vector<float> trace_febex = lisaAnaItem.Get_trace_febex();
+        std::vector<float> trace_febex_0 = lisaAnaItem.Get_trace_febex_0();
+        std::vector<float> trace_MWD = lisaAnaItem.Get_trace_MWD();
 
         
         // ::::: F I L L   H I S T O S :::::
